+++ meta_package_manager/apm.py
@@ -54,7 +54,7 @@ class APM(PackageManager):
         cmd = "{} update --no-confirm".format(self.cli)
         if package_name:
             cmd += " {}".format(package_name)
-        return self.bitbar_cli_format(cmd)
+        return cmd
 
     def update_all_cli(self):
         return self.update_cli()