+++ meta_package_manager/npm.py
@@ -77,7 +77,7 @@ class NPM(PackageManager):
         cmd = "{} -g --progress=false update".format(self.cli)
         if package_name:
             cmd += " {}".format(package_name)
-        return self.bitbar_cli_format(cmd)
+        return cmd
 
     def update_all_cli(self):
         return self.update_cli()