+++ CHANGES.rst
@@ -18,6 +18,8 @@ Changelog
 * Keep a cache of instanciated package manager.
 * Add unittests around package manager definitions.
 * Do not display location of inactive managers, even if hard-coded.
+* Split-up CLI-producing methods and CLI running methods in ``PackageManager``
+  base class.
 
 
 `1.8.0 (2016-08-22) <https://github.com/kdeldycke/meta-package-manager/compare/v1.7.0...v1.8.0>`_