+++ CHANGES.rst
@@ -24,6 +24,7 @@ Changelog
 * Add a new ``sync`` CLI sub-command.
 * Rename managers' ``active`` property to ``available``.
 * Move all package manager definitions in a dedicated folder.
+* Add simple CLI unittests. Closes #2.
 
 
 `1.8.0 (2016-08-22) <https://github.com/kdeldycke/meta-package-manager/compare/v1.7.0...v1.8.0>`_