+++ CHANGES.rst
@@ -25,6 +25,7 @@ Changelog
 * Rename managers' ``active`` property to ``available``.
 * Move all package manager definitions in a dedicated folder.
 * Add simple CLI unittests. Closes #2.
+* Implement ``outdated`` CLI sub-command.
 
 
 `1.8.0 (2016-08-22) <https://github.com/kdeldycke/meta-package-manager/compare/v1.7.0...v1.8.0>`_